+++ Sources/Internal/Platform/TemplateiOS/DPIHelperiOS.mm
@@ -30,7 +30,7 @@
 
 #include "Base/BaseTypes.h"
 #include "Platform/DPIHelper.h"
-#include "DeviceInfo.h"
+#include "Platform/DeviceInfo.h"
 
 #include <UIKit/UIKit.h>
 #include <UIKit/UIDevice.h>
@@ -92,7 +92,7 @@ namespace DAVA
             }
         }
         
-        return outputList.size();
+        return static_cast<uint32>(outputList.size());
     }
     
     uint32 DeterminateExactDPI(List<const DeviceScreenInfo*> &devList)
